requirements.txt
pandas
scikit-learn
datasets
pyyaml
dvc
nltk
#mlflow
transformers
#tourch
accelerate>=1.1.0
seaborn